Update setup script

+++ 3rdParties/src/mysql/setup_ubuntu.sh
@@ -1,10 +1,12 @@
 
-rm libmysqlcppconn*.deb
-wget https://dev.mysql.com/get/Downloads/Connector-C++/libmysqlcppconn7_8.0.21-1ubuntu18.04_amd64.deb
-wget https://dev.mysql.com/get/Downloads/Connector-C++/libmysqlcppconn8-2_8.0.21-1ubuntu18.04_amd64.deb
-wget https://dev.mysql.com/get/Downloads/Connector-C++/libmysqlcppconn-dev_8.0.21-1ubuntu18.04_amd64.deb
+if [ ! -f "libmysqlcppconn-dev_8.0.21-1ubuntu18.04_amd64.deb" ]; then
 
-sudo apt install ./libmysqlcppconn7_8.0.21-1ubuntu18.04_amd64.deb
-sudo apt install ./libmysqlcppconn8-2_8.0.21-1ubuntu18.04_amd64.deb
-sudo apt install ./libmysqlcppconn-dev_8.0.21-1ubuntu18.04_amd64.deb
+	rm libmysqlcppconn*.deb
+	wget https://dev.mysql.com/get/Downloads/Connector-C++/libmysqlcppconn7_8.0.21-1ubuntu18.04_amd64.deb
+	wget https://dev.mysql.com/get/Downloads/Connector-C++/libmysqlcppconn8-2_8.0.21-1ubuntu18.04_amd64.deb
+	wget https://dev.mysql.com/get/Downloads/Connector-C++/libmysqlcppconn-dev_8.0.21-1ubuntu18.04_amd64.deb
 
+	sudo apt install ./libmysqlcppconn7_8.0.21-1ubuntu18.04_amd64.deb
+	sudo apt install ./libmysqlcppconn8-2_8.0.21-1ubuntu18.04_amd64.deb
+	sudo apt install ./libmysqlcppconn-dev_8.0.21-1ubuntu18.04_amd64.deb
+fi
